fix:fix dataste of CHnmr to allow application of dataloader
