+++ molecule_generation/train.py
@@ -5,25 +5,24 @@ import os.path as osp
 import paddle.distributed as dist
 from omegaconf import OmegaConf
 
-from ppmat.datasets import build_dataloader
-from ppmat.datasets import set_signal_handlers
 from ppmat.datasets import CHnmr_dataset
-from ppmat.models.digress.extra_features_graph import DummyExtraFeatures, ExtraFeatures
-from ppmat.metrics.molecular_metrics_discrete import TrainMolecularMetricsDiscrete
-from ppmat.metrics.molecular_metrics import SamplingMolecularMetrics
-from ppmat.utils.visualization import MolecularVisualization
-from ppmat.models.digress.base_model import (
-    MolecularGraphTransformer,
-    ContrastGraphTransformer,
-    ConditionGraphTransformer
-)
+from ppmat.datasets import build_dataloader
 
+# from ppmat.datasets import set_signal_handlers
+from ppmat.metrics.molecular_metrics import SamplingMolecularMetrics
+from ppmat.metrics.molecular_metrics_discrete import TrainMolecularMetricsDiscrete
+from ppmat.models.digress.base_model import ConditionGraphTransformer
+from ppmat.models.digress.base_model import ContrastGraphTransformer
+from ppmat.models.digress.base_model import MolecularGraphTransformer
+from ppmat.models.digress.extra_features_graph import DummyExtraFeatures
+from ppmat.models.digress.extra_features_graph import ExtraFeatures
 from ppmat.optimizer import build_optimizer
 from ppmat.trainer.trainer_multimodal import TrainerCLIP
 from ppmat.trainer.trainer_multimodal import TrainerGraph
 from ppmat.trainer.trainer_multimodal import TrainerMultiModal
 from ppmat.utils import logger
 from ppmat.utils import misc
+from ppmat.utils.visualization import MolecularVisualization
 
 if dist.get_world_size() > 1:
     dist.fleet.init(is_collective=True)
@@ -64,37 +63,62 @@ if __name__ == "__main__":
     logger.info(f"Set random seed to {seed}")
 
     # load dataloader from config
-    #train_data_cfg = config["Dataset"]["train"]
-    #train_loader = build_dataloader(train_data_cfg)
-    #val_data_cfg = config["Dataset"]["val"]
-    #val_loader = build_dataloader(val_data_cfg)
-    #test_data_cfg = config["Dataset"]["test"]
-    #test_loader = build_dataloader(test_data_cfg)
+    train_data_cfg = config["Dataset"]["train"]
+    train_loader = build_dataloader(train_data_cfg)
+    val_data_cfg = config["Dataset"]["val"]
+    val_loader = build_dataloader(val_data_cfg)
+    test_data_cfg = config["Dataset"]["test"]
+    test_loader = build_dataloader(test_data_cfg)
+
+    for i, batch in enumerate(train_loader()):
+        print(i)
+        print(type(batch))
+        print(batch)
+        break
 
     # build datasetinfo
-    #dataset_infos = CHnmr_dataset.CHnmrinfos(datamodule=train_loader, cfg=config)
-    #train_smiles = CHnmr_dataset.get_train_smiles(cfg=config, train_dataloader=train_loader,
-    #                                                    dataset_infos=dataset_infos, evaluate_dataset=False)
-    
+    dataset_infos = CHnmr_dataset.CHnmrinfos(datamodule=train_loader, cfg=config)
+    train_smiles = CHnmr_dataset.get_train_smiles(
+        cfg=config,
+        train_dataloader=train_loader,
+        dataset_infos=dataset_infos,
+        evaluate_dataset=False,
+    )
+
     # extra features
     if config["Model"]["model_setting"]["extra_features"] is not None:
-        extra_features = ExtraFeatures(config["Model"]["model_setting"]["extra_features"], dataset_info=dataset_infos)
-        domain_features = ExtraFeatures(config["Model"]["model_setting"]["extra_features"], dataset_infos=dataset_infos)
+        extra_features = ExtraFeatures(
+            config["Model"]["model_setting"]["extra_features"],
+            dataset_info=dataset_infos,
+        )
+        domain_features = ExtraFeatures(
+            config["Model"]["model_setting"]["extra_features"],
+            dataset_infos=dataset_infos,
+        )
     else:
         extra_features = DummyExtraFeatures()
         domain_features = DummyExtraFeatures()
-    
-    #dataset_infos.compute_input_output_dims(datamodule=train_loader, extra_features=extra_features, domain_features=domain_features, conditionDim=config.model.model_setting.conditdim)
+
+    dataset_infos.compute_input_output_dims(
+        datamodule=train_loader,
+        extra_features=extra_features,
+        domain_features=domain_features,
+        conditionDim=config.model.model_setting.conditdim,
+    )
     train_metrics = TrainMolecularMetricsDiscrete(dataset_infos)
     sampling_metrics = SamplingMolecularMetrics(dataset_infos, train_smiles)
-    visualization_tools = MolecularVisualization(config.Dataset.train.remove_h, dataset_infos=dataset_infos)
-    
-    model_kwargs = {'dataset_infos': dataset_infos,
-                    'train_metrics': train_metrics,
-                    'sampling_metrics': sampling_metrics,
-                    'visualization_tools': visualization_tools,
-                    'extra_features': extra_features,
-                    'domain_features': domain_features}
+    visualization_tools = MolecularVisualization(
+        config.Dataset.train.remove_h, dataset_infos=dataset_infos
+    )
+
+    model_kwargs = {
+        "dataset_infos": dataset_infos,
+        "train_metrics": train_metrics,
+        "sampling_metrics": sampling_metrics,
+        "visualization_tools": visualization_tools,
+        "extra_features": extra_features,
+        "domain_features": domain_features,
+    }
 
     # initialize trainer
     if args.step == 1:
